requirements.txt
django>=1.7,<1.8
django-html5-colorfield==1.0
django-autoslug==1.7.2
